Renaming DB Models Folder Using CamelCase Convention
