+++ backend/.sequelizerc
@@ -2,7 +2,7 @@ const path = require('path')
 
 module.exports = {
   'config': path.resolve('config', 'config.json'),
-  'models-path': path.resolve('models', 'db_models'),
+  'models-path': path.resolve('models', 'dbModels'),
   'seeders-path': path.resolve('seeders'),
   'migrations-path': path.resolve('migrations')
 };
